edge/data-injector/scripts/requirements.txt
wheel
mqtt_client
requests
paho.mqtt